+++ test/unit/test_counter.dart
@@ -46,7 +46,8 @@ abstract class Counter implements Built<Counter, CounterBuilder> {
 // Middleware
 
 abstract class MiddlewareActions extends ReduxActions {
-  ActionDispatcher<int> increment;
+  ActionDispatcher<int> doubleIt;
+  ActionDispatcher<int> tripleIt;
 
   MiddlewareActions._();
   factory MiddlewareActions() => new _$MiddlewareActions();
@@ -54,12 +55,23 @@ abstract class MiddlewareActions extends ReduxActions {
 
 var counterMiddleware =
     (new MiddlewareBuilder<Counter, CounterBuilder, CounterActions>()
-          ..add(MiddlewareActionsNames.increment, _doubleIt))
+          ..add(MiddlewareActionsNames.doubleIt, _doubleIt)
+          ..combine(tripleItMiddlewareBuilder))
         .build();
 
 void _doubleIt(MiddlewareApi<Counter, CounterBuilder, CounterActions> api,
     ActionHandler next, Action<int> action) {
   api.actions.increment(api.state.count * 2);
+  next(action);
+}
+
+var tripleItMiddlewareBuilder =
+    new MiddlewareBuilder<Counter, CounterBuilder, CounterActions>()
+      ..add(MiddlewareActionsNames.tripleIt, _tripleIt);
+
+void _tripleIt(MiddlewareApi<Counter, CounterBuilder, CounterActions> api,
+    ActionHandler next, Action<int> action) {
+  api.actions.increment(api.state.count * 3);
   next(action);
 }